+++ src/GHCi/DAP/IFData.hs
@@ -10,7 +10,42 @@ Implementation of DAP interface data type.
 @see : https://github.com/Microsoft/vscode-debugadapter-node/blob/master/protocol/src/debugProtocol.ts
 
 -}
-module GHCi.DAP.IFData where
+module GHCi.DAP.IFData (
+    -- * setBreakpoints
+    SetBreakpointsArguments(..)
+  , SourceBreakpoint(..)
+  , SetBreakpointsResponseBody(..)
+  , Breakpoint(..)
+  , defaultBreakpoint
+    -- * continue
+  , ContinueArguments(..)
+  , StoppedEventBody(..)
+  , defaultStoppedEventBody
+    -- * scopes
+  , ScopesArguments(..)
+  , ScopesBody(..)
+  , Scope(..)
+  , defaultScope
+    -- * stackTrace
+  , StackTraceArguments(..)
+  , StackTraceBody(..)
+  , defaultStackTraceBody
+  , StackFrame(..)
+  , defaultStackFrame
+    -- * variables
+  , VariablesArguments(..)
+  , VariablesBody(..)
+  , Variable(..)
+  , defaultVariable
+    -- * evaluate
+  , EvaluateArguments(..)
+  , EvaluateBody(..)
+  , defaultEvaluateBody
+    -- * commons
+  , VariablePresentationHint(..)
+  , Source(..)
+  , defaultSource
+) where
 
 
 -- |
@@ -18,7 +53,7 @@ module GHCi.DAP.IFData where
 --
 data VariablesArguments =
   VariablesArguments {
-    variablesReferenceVariablesArguments :: Int  -- The Variable reference.
+    variablesReferenceVariablesArguments :: Int  -- ^The Variable reference.
   } deriving (Show, Read, Eq)
 
 
@@ -27,7 +62,7 @@ data VariablesArguments =
 --
 data VariablesBody =
   VariablesBody {
-    variablesVariablesBody :: [Variable]  -- All (or a range) of variables for the given variable reference.
+    variablesVariablesBody :: [Variable]  -- ^All (or a range) of variables for the given variable reference.
   } deriving (Show, Read, Eq)
 
 
@@ -124,8 +159,8 @@ data VariablePresentationHint =
 --
 data EvaluateArguments =
   EvaluateArguments {
-    expressionEvaluateArguments :: String     -- The expression to evaluate. 
-  , frameIdEvaluateArguments    :: Maybe Int  -- Evaluate the expression in the scope of this stack frame. If not specified, the expression is evaluated in the global scope. 
+    expressionEvaluateArguments :: String     -- ^The expression to evaluate. 
+  , frameIdEvaluateArguments    :: Maybe Int  -- ^Evaluate the expression in the scope of this stack frame. If not specified, the expression is evaluated in the global scope. 
 
   {-|
     The context in which the evaluate request is run.
@@ -173,7 +208,7 @@ defaultEvaluateBody = EvaluateBody {
 --
 data ScopesArguments =
   ScopesArguments {
-    frameIdScopesArguments :: Int  -- Retrieve the scopes for this stackframe.
+    frameIdScopesArguments :: Int  -- ^Retrieve the scopes for this stackframe.
   } deriving (Show, Read, Eq)
 
 
@@ -215,9 +250,8 @@ defaultScope = Scope {
 --
 data SetBreakpointsArguments =
   SetBreakpointsArguments {
-    sourceSetBreakpointsArguments         :: Source              --  The source location of the breakpoints; either source.path or source.reference must be specified. 
-  , breakpointsSetBreakpointsArguments    :: [SourceBreakpoint]  -- The code locations of the breakpoints.
-  --, sourceModifiedSetBreakpointsArguments :: Bool              --  A value of true indicates that the underlying source has been modified which results in new breakpoint locations. 
+    sourceSetBreakpointsArguments         :: Source              -- ^The source location of the breakpoints; either source.path or source.reference must be specified. 
+  , breakpointsSetBreakpointsArguments    :: [SourceBreakpoint]  -- ^The code locations of the breakpoints.
   } deriving (Show, Read, Eq)
 
 
@@ -226,14 +260,10 @@ data SetBreakpointsArguments =
 --
 data Source =
   Source {
-    nameSource             :: Maybe String  -- The short name of the source. Every source returned from the debug adapter has a name. When sending a source to the debug adapter this name is optional.
-  , pathSource             :: String        -- The path of the source to be shown in the UI. It is only used to locate and load the content of the source if no sourceReference is specified (or its vaule is 0).
-  , sourceReferenceSource  :: Maybe Int     -- If sourceReference > 0 the contents of the source must be retrieved through the SourceRequest (even if a path is specified). A sourceReference is only valid for a session, so it must not be used to persist a source.
-  -- , presentationHintSource :: Maybe String  -- An optional hint for how to present the source in the UI. A value of 'deemphasize' can be used to indicate that the source is not available or that it is skipped on stepping.
-  , origineSource          :: Maybe String  -- The (optional) origin of this source: possible values 'internal module', 'inlined content from source map', etc.
-  -- , sourcesSource          :: [Source]      -- An optional list of sources that are related to this source. These may be the source that generated this source.
-  -- , adapterDataSource      :: Maybe String  -- Optional data that a debug adapter might want to loop through the client. The client should leave the data intact and persist it across sessions. The client should not interpret the data.
-  -- , checksumsSource        :: []            -- The checksums associated with this file.
+    nameSource             :: Maybe String  -- ^The short name of the source. Every source returned from the debug adapter has a name. When sending a source to the debug adapter this name is optional.
+  , pathSource             :: String        -- ^The path of the source to be shown in the UI. It is only used to locate and load the content of the source if no sourceReference is specified (or its vaule is 0).
+  , sourceReferenceSource  :: Maybe Int     -- ^If sourceReference > 0 the contents of the source must be retrieved through the SourceRequest (even if a path is specified). A sourceReference is only valid for a session, so it must not be used to persist a source.
+  , origineSource          :: Maybe String     -- ^The (optional) origin of this source: possible values 'internal module', 'inlined content from source map', etc.
   } deriving (Show, Read, Eq)
 
 
@@ -244,11 +274,7 @@ defaultSource = Source {
     nameSource             = Nothing
   , pathSource             = ""
   , sourceReferenceSource  = Nothing
---  , presentationHintSource = Nothing
   , origineSource          = Nothing
---  , sourcesSource          = []
---  , adapterDataSource      = Nothing
---  , checksumsSource        = []
   }
 
 
@@ -257,10 +283,10 @@ defaultSource = Source {
 --
 data SourceBreakpoint =
   SourceBreakpoint {
-    lineSourceBreakpoint         :: Int           -- The source line of the breakpoint.
-  , columnSourceBreakpoint       :: Maybe Int     -- An optional source column of the breakpoint. 
-  , conditionSourceBreakpoint    :: Maybe String  -- An optional expression for conditional breakpoints.
-  , hitConditionSourceBreakpoint :: Maybe String  -- An optional expression that controls how many hits of the breakpoint are ignored. The backend is expected to interpret the expression as needed.
+    lineSourceBreakpoint         :: Int           -- ^The source line of the breakpoint.
+  , columnSourceBreakpoint       :: Maybe Int     -- ^An optional source column of the breakpoint. 
+  , conditionSourceBreakpoint    :: Maybe String  -- ^An optional expression for conditional breakpoints.
+  , hitConditionSourceBreakpoint :: Maybe String  -- ^An optional expression that controls how many hits of the breakpoint are ignored. The backend is expected to interpret the expression as needed.
   } deriving (Show, Read, Eq)
 
 
@@ -273,7 +299,7 @@ data SourceBreakpoint =
 --
 data SetBreakpointsResponseBody =
   SetBreakpointsResponseBody {
-    breakpointsSetBreakpointsResponseBody :: [Breakpoint]  -- Information about the breakpoints. The array elements are in the same order as the elements of the 'breakpoints' (or the deprecated 'lines') in the SetBreakpointsArguments.
+    breakpointsSetBreakpointsResponseBody :: [Breakpoint]  -- ^Information about the breakpoints. The array elements are in the same order as the elements of the 'breakpoints' (or the deprecated 'lines') in the SetBreakpointsArguments.
   } deriving (Show, Read, Eq)
 
 
@@ -282,14 +308,14 @@ data SetBreakpointsResponseBody =
 --
 data Breakpoint =
   Breakpoint {
-    idBreakpoint        :: Maybe Int -- An optional unique identifier for the breakpoint.
-  , verifiedBreakpoint  :: Bool      -- If true breakpoint could be set (but not necessarily at the desired location).
-  , messageBreakpoint   :: String    -- An optional message about the state of the breakpoint. This is shown to the user and can be used to explain why a breakpoint could not be verified.
-  , sourceBreakpoint    :: Source    -- The source where the breakpoint is located.
-  , lineBreakpoint      :: Int       -- The start line of the actual range covered by the breakpoint.
-  , columnBreakpoint    :: Int       -- An optional start column of the actual range covered by the breakpoint.
-  , endLineBreakpoint   :: Int       -- An optional end line of the actual range covered by the breakpoint.
-  , endColumnBreakpoint :: Int       -- An optional end column of the actual range covered by the breakpoint. If no end line is given, then the end column is assumed to be in the start line.
+    idBreakpoint        :: Maybe Int -- ^An optional unique identifier for the breakpoint.
+  , verifiedBreakpoint  :: Bool      -- ^If true breakpoint could be set (but not necessarily at the desired location).
+  , messageBreakpoint   :: String    -- ^An optional message about the state of the breakpoint. This is shown to the user and can be used to explain why a breakpoint could not be verified.
+  , sourceBreakpoint    :: Source    -- ^The source where the breakpoint is located.
+  , lineBreakpoint      :: Int       -- ^The start line of the actual range covered by the breakpoint.
+  , columnBreakpoint    :: Int       -- ^An optional start column of the actual range covered by the breakpoint.
+  , endLineBreakpoint   :: Int       -- ^An optional end line of the actual range covered by the breakpoint.
+  , endColumnBreakpoint :: Int       -- ^An optional end column of the actual range covered by the breakpoint. If no end line is given, then the end column is assumed to be in the start line.
   } deriving (Show, Read, Eq)
 
 
@@ -313,8 +339,8 @@ defaultBreakpoint = Breakpoint {
 --
 data ContinueArguments =
   ContinueArguments {
-    threadIdContinueArguments :: Int --  Continue execution for the specified thread (if possible). If the backend cannot continue on a single thread but will continue on all threads, it should set the allThreadsContinued attribute in the response to true.
-  , exprContinueArguments     :: Maybe String -- ADD: haskell-dap
+    threadIdContinueArguments :: Int          -- ^Continue execution for the specified thread (if possible). If the backend cannot continue on a single thread but will continue on all threads, it should set the allThreadsContinued attribute in the response to true.
+  , exprContinueArguments     :: Maybe String -- ^ADD: haskell-dap
   } deriving (Show, Read, Eq)
 
 
@@ -325,14 +351,16 @@ data ContinueArguments =
 --
 data StoppedEventBody =
   StoppedEventBody {
-    reasonStoppedEventBody            :: String  -- The reason for the event.For backward compatibility this string is shown in the UI if the 'description' attribute is missing (but it must not be translated).Values: 'step', 'breakpoint', 'exception', 'pause', 'entry', etc.
-  , descriptionStoppedEventBody       :: String  -- The full reason for the event, e.g. 'Paused on exception'. This string is shown in the UI as is.
-  , threadIdStoppedEventBody          :: Int     -- The thread which was stopped.
-  , textStoppedEventBody              :: String  -- Additional information. E.g. if reason is 'exception', text contains the exception name. This string is shown in the UI. 
+    reasonStoppedEventBody            :: String  -- ^The reason for the event.For backward compatibility this string is shown in the UI if the 'description' attribute is missing (but it must not be translated).Values: 'step', 'breakpoint', 'exception', 'pause', 'entry', etc.
+  , descriptionStoppedEventBody       :: String  -- ^The full reason for the event, e.g. 'Paused on exception'. This string is shown in the UI as is.
+  , threadIdStoppedEventBody          :: Int     -- ^The thread which was stopped.
+  , textStoppedEventBody              :: String  -- ^Additional information. E.g. if reason is 'exception', text contains the exception name. This string is shown in the UI. 
 
-  -- If allThreadsStopped is true, a debug adapter can announce that all threads have stopped.
-  -- The client should use this information to enable that all threads can be expanded to access their stacktraces.
-  -- If the attribute is missing or false, only the thread with the given threadId can be expanded.
+  {-|
+     If allThreadsStopped is true, a debug adapter can announce that all threads have stopped.
+     The client should use this information to enable that all threads can be expanded to access their stacktraces.
+     If the attribute is missing or false, only the thread with the given threadId can be expanded.
+  -}
   , allThreadsStoppedStoppedEventBody :: Bool
   } deriving (Show, Read, Eq)
 
@@ -353,9 +381,9 @@ defaultStoppedEventBody = StoppedEventBody {
 --
 data StackTraceArguments =
   StackTraceArguments {
-    threadIdStackTraceArguments   :: Int        -- Retrieve the stacktrace for this thread.
-  , startFrameStackTraceArguments :: Maybe Int  -- The index of the first frame to return; if omitted frames start at 0.
-  , levelsStackTraceArguments     :: Int        -- The maximum number of frames to return. If levels is not specified or 0, all frames are returned.
+    threadIdStackTraceArguments   :: Int        -- ^Retrieve the stacktrace for this thread.
+  , startFrameStackTraceArguments :: Maybe Int  -- ^The index of the first frame to return; if omitted frames start at 0.
+  , levelsStackTraceArguments     :: Int        -- ^The maximum number of frames to return. If levels is not specified or 0, all frames are returned.
   } deriving (Show, Read, Eq)
 
 
@@ -364,8 +392,8 @@ data StackTraceArguments =
 --
 data StackTraceBody =
   StackTraceBody {
-    stackFramesStackTraceBody :: [StackFrame]  -- The frames of the stackframe. If the array has length zero, there are no stackframes available. This means that there is no location information available.
-  , totalFramesStackTraceBody :: Int           -- The total number of frames available. 
+    stackFramesStackTraceBody :: [StackFrame]  -- ^The frames of the stackframe. If the array has length zero, there are no stackframes available. This means that there is no location information available.
+  , totalFramesStackTraceBody :: Int           -- ^The total number of frames available. 
   } deriving (Show, Read, Eq)
 
 -- |
@@ -381,13 +409,13 @@ defaultStackTraceBody = StackTraceBody {
 --
 data StackFrame =
   StackFrame {
-    idStackFrame        :: Int     -- An identifier for the stack frame. It must be unique across all threads. This id can be used to retrieve the scopes of the frame with the 'scopesRequest' or to restart the execution of a stackframe.
-  , nameStackFrame      :: String  -- The name of the stack frame, typically a method name.
-  , sourceStackFrame    :: Source  -- The optional source of the frame.
-  , lineStackFrame      :: Int     -- The line within the file of the frame. If source is null or doesn't exist, line is 0 and must be ignored.
-  , columnStackFrame    :: Int     -- The column within the line. If source is null or doesn't exist, column is 0 and must be ignored. 
-  , endLineStackFrame   :: Int     -- An optional end line of the range covered by the stack frame.
-  , endColumnStackFrame :: Int     -- An optional end column of the range covered by the stack frame.
+    idStackFrame        :: Int     -- ^An identifier for the stack frame. It must be unique across all threads. This id can be used to retrieve the scopes of the frame with the 'scopesRequest' or to restart the execution of a stackframe.
+  , nameStackFrame      :: String  -- ^The name of the stack frame, typically a method name.
+  , sourceStackFrame    :: Source  -- ^The optional source of the frame.
+  , lineStackFrame      :: Int     -- ^The line within the file of the frame. If source is null or doesn't exist, line is 0 and must be ignored.
+  , columnStackFrame    :: Int     -- ^The column within the line. If source is null or doesn't exist, column is 0 and must be ignored. 
+  , endLineStackFrame   :: Int     -- ^An optional end line of the range covered by the stack frame.
+  , endColumnStackFrame :: Int     -- ^An optional end column of the range covered by the stack frame.
   } deriving (Show, Read, Eq)